+++ tests/setup/index.ts
@@ -11,13 +11,30 @@ global.crypto = {
 // Jest environment not contains DataTransfer object, so mock a DataTransfer class
 // @ts-ignore
 global.DataTransfer = class DataTransfer {
-  clearData() {}
+  clearData() { }
   getData(type: string) {
     if (type === 'text/plain') return ''
     return []
   }
-  setData() {}
+  setData() { }
   get files() {
     return [new File(['124'], 'test.jpg')]
+  }
+}
+
+
+global.ResizeObserver = class ResizeObserver {
+  constructor(callback) {
+    // @ts-ignore
+    this.callback = callback;
+  }
+  observe() {
+    // 可以根据需要添加具体实现
+  }
+  unobserve() {
+    // 可以根据需要添加具体实现
+  }
+  disconnect() {
+    // 可以根据需要添加具体实现
   }
 }